Adding a backend on layout settings to handle widgets that modify xkb-options
An Xkb modifier is tied to several xkb-options, where every option has the command
on xkb-options that will enable it and a description. A modifier that has a group
of different commands will only set the active one on xkb-options, which means 
that setting two options of the same group at the same time should have no meaning
at all, and should not be allowed.