Adding handlers to all Layout page widgets
Each widget that changes xkb-options has one Xkb_modifier asigned to it, when a
property can be changed for multiple layouts, only one modifier should be used
so that it's only changed once and kept in sync. For this a named modifier is
used, and it's retrieved using settings.get_xkb_modifier_by_name ().